♻️ refactor: move connection/channel/exchange parsers into their respective files, remove unused message_types.go and tidy imports
