--- internal/core/amqp/message_types.go
@@ -1,1 +1,0 @@
-package amqp
