+++ internal/core/amqp/parsers.go
@@ -1,7 +1,6 @@
 package amqp
 
 import (
-	"bytes"
 	"encoding/binary"
 	"fmt"
 
@@ -218,375 +217,3 @@ func parseBodyFrame(payloadSize uint32, payload []byte) (*ChannelState, error) {
 	return state, nil
 
 }
-
-// REGION Exchange_Methods
-func parseExchangeMethod(methodID uint16, payload []byte) (interface{}, error) {
-	switch methodID {
-	case uint16(EXCHANGE_DECLARE):
-		log.Debug().Msg("Received EXCHANGE_DECLARE frame \n")
-		return parseExchangeDeclareFrame(payload)
-	case uint16(EXCHANGE_DELETE):
-		log.Debug().Msg("Received EXCHANGE_DELETE frame \n")
-		return parseExchangeDeleteFrame(payload)
-
-	default:
-		return nil, fmt.Errorf("unknown method ID: %d", methodID)
-	}
-}
-
-// REGION Connection_Methods
-
-func parseConnectionMethod(methodID uint16, payload []byte) (any, error) {
-	switch methodID {
-	case uint16(CONNECTION_START):
-		log.Debug().Msg("Received CONNECTION_START frame")
-		return parseConnectionStartFrame(payload)
-
-	case uint16(CONNECTION_START_OK):
-		log.Trace().Bytes("payload", payload).Msg("Received connection.start-ok")
-		return parseConnectionStartOkFrame(payload)
-
-	case uint16(CONNECTION_TUNE):
-		log.Debug().Msg("Received CONNECTION_TUNE frame")
-		tuneResponse, err := parseConnectionTuneFrame(payload)
-		if err != nil {
-			return nil, fmt.Errorf("failed to parse connection.tune frame: %v", err)
-		}
-		tuneOkRequest := createConnectionTuneOkFrame(fineTune(tuneResponse))
-		return tuneOkRequest, nil
-
-	case uint16(CONNECTION_TUNE_OK):
-		log.Debug().Msg("Received CONNECTION_TUNE_OK frame")
-		return parseConnectionTuneOkFrame(payload)
-
-	case uint16(CONNECTION_OPEN):
-		log.Debug().Msg("Received CONNECTION_OPEN frame")
-		return parseConnectionOpenFrame(payload)
-
-	case uint16(CONNECTION_OPEN_OK):
-		log.Debug().Msg("Received CONNECTION_OPEN_OK frame")
-		return parseConnectionOpenOkFrame(payload)
-
-	case uint16(CONNECTION_CLOSE):
-		request, err := parseConnectionCloseFrame(payload)
-		if err != nil {
-			return nil, fmt.Errorf("failed to parse connection.close frame: %v", err)
-		}
-		request.MethodID = methodID
-		content, ok := request.Content.(*ConnectionCloseMessage)
-		if !ok {
-			return nil, fmt.Errorf("invalid message content type")
-		}
-		log.Debug().Msgf("Received CONNECTION_CLOSE: (%d) '%s'", content.ReplyCode, content.ReplyText)
-		return request, nil
-
-	case uint16(CONNECTION_CLOSE_OK):
-		request, err := parseConnectionCloseOkFrame(payload)
-		if err != nil {
-			return nil, fmt.Errorf("failed to parse CONNECTION_CLOSE_OK frame: %v", err)
-		}
-		request.MethodID = methodID
-		log.Debug().Msg("Received CONNECTION_CLOSE_OK")
-		return request, nil
-
-	default:
-		return nil, fmt.Errorf("unknown method ID: %d", methodID)
-	}
-}
-
-func parseConnectionCloseFrame(payload []byte) (*RequestMethodMessage, error) {
-	if len(payload) < 12 {
-		return nil, fmt.Errorf("frame too short")
-	}
-
-	index := uint16(0)
-	replyCode := binary.BigEndian.Uint16(payload[index : index+2])
-	index += 2
-	replyTextLen := payload[index]
-	index++
-	replyText := string(payload[index : index+uint16(replyTextLen)])
-	index += uint16(replyTextLen)
-	classID := binary.BigEndian.Uint16(payload[index : index+2])
-	index += 2
-	methodID := binary.BigEndian.Uint16(payload[index : index+2])
-	msg := &ConnectionCloseMessage{
-		ReplyCode: replyCode,
-		ReplyText: replyText,
-		ClassID:   classID,
-		MethodID:  methodID,
-	}
-	request := &RequestMethodMessage{
-		Content: msg,
-	}
-	log.Trace().Msgf("connection close request: %+v\n", request)
-	return request, nil
-}
-
-func parseConnectionCloseOkFrame(payload []byte) (*RequestMethodMessage, error) {
-	return &RequestMethodMessage{Content: payload}, nil
-}
-
-func parseConnectionOpenFrame(payload []byte) (interface{}, error) {
-	if len(payload) < 1 {
-		return nil, fmt.Errorf("payload too short")
-	}
-	buf := bytes.NewReader(payload)
-	virtualHost, err := DecodeShortStr(buf)
-	if err != nil {
-		return nil, fmt.Errorf("failed to decode virtual host: %v", err)
-	}
-	return &ConnectionOpen{
-		VirtualHost: virtualHost,
-	}, nil
-}
-
-func parseConnectionOpenOkFrame(payload []byte) (interface{}, error) {
-	if len(payload) < 1 {
-		return nil, fmt.Errorf("payload too short")
-	}
-	// buf := bytes.NewReader(payload)
-	return nil, nil
-}
-
-func parseConnectionTuneFrame(payload []byte) (*ConnectionTune, error) {
-	if len(payload) < 6 {
-		return nil, fmt.Errorf("payload too short")
-	}
-	buf := bytes.NewReader(payload)
-	channelMax, err := DecodeShortInt(buf)
-	if err != nil {
-		return nil, fmt.Errorf("failed to decode channel max: %v", err)
-	}
-
-	frameMax, err := DecodeLongInt(buf)
-	if err != nil {
-		return nil, fmt.Errorf("failed to decode frame max: %v", err)
-	}
-	heartbeat, err := DecodeShortInt(buf)
-	if err != nil {
-		return nil, fmt.Errorf("failed to decode heartbeat: %v", err)
-	}
-	return &ConnectionTune{
-		ChannelMax: channelMax,
-		FrameMax:   frameMax,
-		Heartbeat:  heartbeat,
-	}, nil
-}
-
-func parseConnectionTuneOkFrame(payload []byte) (interface{}, error) {
-	if len(payload) < 6 {
-		return nil, fmt.Errorf("payload too short")
-	}
-	buf := bytes.NewReader(payload)
-	channelMax, err := DecodeShortInt(buf)
-	if err != nil {
-		return nil, fmt.Errorf("failed to decode channel max: %v", err)
-	}
-
-	frameMax, err := DecodeLongInt(buf)
-	if err != nil {
-		return nil, fmt.Errorf("failed to decode frame max: %v", err)
-	}
-	heartbeat, err := DecodeShortInt(buf)
-	if err != nil {
-		return nil, fmt.Errorf("failed to decode heartbeat: %v", err)
-	}
-	return &ConnectionTune{
-		ChannelMax: channelMax,
-		FrameMax:   frameMax,
-		Heartbeat:  heartbeat,
-	}, nil
-}
-
-func parseConnectionStartFrame(payload []byte) (*ConnectionStartFrame, error) {
-	if len(payload) < 6 {
-		return nil, fmt.Errorf("payload too short")
-	}
-
-	buf := bytes.NewReader(payload)
-
-	// Decode version-major and version-minor
-	versionMajor, err := buf.ReadByte()
-	if err != nil {
-		return nil, fmt.Errorf("failed to decode version-major: %v", err)
-	}
-	log.Trace().Msgf("Version Major: %d\n", versionMajor)
-
-	versionMinor, err := buf.ReadByte()
-	if err != nil {
-		return nil, fmt.Errorf("failed to decode version-minor: %v", err)
-	}
-	log.Trace().Msgf("Version Minor: %d\n", versionMinor)
-
-	/***Server Properties*/
-	serverPropertiesStr, err := DecodeLongStr(buf)
-	if err != nil {
-		return nil, fmt.Errorf("failed to decode client properties: %v", err)
-	}
-	serverProperties, err := DecodeTable([]byte(serverPropertiesStr))
-	if err != nil {
-		return nil, fmt.Errorf("failed to decode client properties: %v", err)
-	}
-	log.Trace().Msgf("Server Properties: %+v\n", serverProperties)
-
-	/***Mechanisms*/
-	mechanismsStr, err := DecodeLongStr(buf)
-	if err != nil {
-		return nil, fmt.Errorf("failed to decode mechanism: %v", err)
-	}
-	log.Trace().Msgf("Mechanisms: %s\n", mechanismsStr)
-
-	/***Locales*/
-	localesStr, err := DecodeLongStr(buf)
-	if err != nil {
-		return nil, fmt.Errorf("failed to decode locale: %v", err)
-	}
-	log.Trace().Msgf("Locales: %s\n", localesStr)
-
-	return &ConnectionStartFrame{
-		VersionMajor:     versionMajor,
-		VersionMinor:     versionMinor,
-		ServerProperties: serverProperties,
-		Mechanisms:       mechanismsStr,
-		Locales:          localesStr,
-	}, nil
-}
-
-func parseConnectionStartOkFrame(payload []byte) (*ConnectionStartOk, error) {
-	if len(payload) < 6 {
-		return nil, fmt.Errorf("payload too short")
-	}
-
-	buf := bytes.NewReader(payload)
-
-	clientPropertiesStr, err := DecodeLongStr(buf)
-	if err != nil {
-		return nil, fmt.Errorf("failed to decode client properties: %v", err)
-	}
-
-	clientProperties, err := DecodeTable([]byte(clientPropertiesStr))
-	if err != nil {
-		return nil, fmt.Errorf("failed to decode client properties: %v", err)
-	}
-
-	mechanism, err := DecodeShortStr(buf)
-	if err != nil {
-		return nil, fmt.Errorf("failed to decode mechanism: %v", err)
-	}
-
-	if mechanism != "PLAIN" {
-		return nil, fmt.Errorf("unsupported mechanism: %s", mechanism)
-	}
-
-	security, err := DecodeSecurityPlain(buf)
-	if err != nil {
-		return nil, fmt.Errorf("failed to decode security: %v", err)
-	}
-
-	locale, err := DecodeShortStr(buf)
-	if err != nil {
-		return nil, fmt.Errorf("failed to decode locale: %v", err)
-	}
-
-	return &ConnectionStartOk{
-		ClientProperties: clientProperties,
-		Mechanism:        mechanism,
-		Response:         security,
-		Locale:           locale,
-	}, nil
-}
-
-// REGION Channel_Methods
-func parseChannelMethod(methodID uint16, payload []byte) (interface{}, error) {
-	switch methodID {
-	case uint16(CHANNEL_OPEN):
-		log.Debug().Msg("Received CHANNEL_OPEN frame")
-		return parseChannelOpenFrame(payload)
-
-	case uint16(CHANNEL_OPEN_OK):
-		log.Debug().Msg("Received CHANNEL_OPEN_OK frame")
-		return parseChannelOpenOkFrame(payload)
-
-	case uint16(CHANNEL_CLOSE):
-		log.Debug().Msg("Received CHANNEL_CLOSE frame")
-		return parseChannelCloseFrame(payload)
-
-	case uint16(CHANNEL_CLOSE_OK):
-		log.Debug().Msg("Received CHANNEL_CLOSE_OK frame")
-		return parseChannelCloseOkFrame(payload)
-
-	default:
-		return nil, fmt.Errorf("unknown method ID: %d", methodID)
-	}
-}
-
-func parseChannelOpenFrame(payload []byte) (*RequestMethodMessage, error) {
-	if len(payload) < 1 {
-		return nil, fmt.Errorf("payload too short")
-	}
-	request := &RequestMethodMessage{
-		Content: nil,
-	}
-
-	return request, nil
-}
-
-func parseChannelOpenOkFrame(payload []byte) (*RequestMethodMessage, error) {
-	if len(payload) < 1 {
-		return nil, fmt.Errorf("payload too short")
-	}
-	request := &RequestMethodMessage{
-		Content: nil,
-	}
-
-	return request, nil
-}
-
-func parseChannelCloseFrame(payload []byte) (*RequestMethodMessage, error) {
-	log.Trace().Msgf("Received CHANNEL_CLOSE frame: %x \n", payload)
-	if len(payload) < 6 {
-		return nil, fmt.Errorf("frame too short")
-	}
-
-	buf := bytes.NewReader(payload)
-	replyCode, err := DecodeShortInt(buf)
-	if err != nil {
-		return nil, fmt.Errorf("failed decoding reply code: %v", err)
-	}
-	replyText, err := DecodeShortStr(buf)
-	if err != nil {
-		return nil, fmt.Errorf("failed decoding reply text: %v", err)
-	}
-	classID, err := DecodeShortInt(buf)
-	if err != nil {
-		return nil, fmt.Errorf("failed decoding class id: %v", err)
-	}
-	methodID, err := DecodeShortInt(buf)
-	if err != nil {
-		return nil, fmt.Errorf("failed decoding method id: %v", err)
-	}
-
-	msg := &ChannelCloseMessage{
-		ReplyCode: replyCode,
-		ReplyText: replyText,
-		ClassID:   classID,
-		MethodID:  methodID,
-	}
-	request := &RequestMethodMessage{
-		Content: msg,
-	}
-	return request, nil
-}
-
-func parseChannelCloseOkFrame(payload []byte) (*RequestMethodMessage, error) {
-	if len(payload) > 1 {
-		return nil, fmt.Errorf("unexxpected payload length")
-	}
-	request := &RequestMethodMessage{
-		Content: nil,
-	}
-
-	return request, nil
-}
-
-// REGION Basic_headers
